提交修改

+++ src/main/java/com/xinkao/erp/exercise/param/PanJuanParam.java
@@ -38,7 +38,7 @@ public class PanJuanParam {
     private BigDecimal coverageRate;
 
     @ApiModelProperty("题目操作时长（单位：秒）")
-    private Integer operationDuration;
+    private Long operationDuration;
 
     @ApiModelProperty("用户得分")
     private Integer score;
